fixed naming error in windows due to datetime format
in windows certain characters are not allowed to be in naming scheme including symbol colon. since we are saving the image with its datetime which contain colon symbol thus windows throws invalid name error. so to fix this, the " : " in the name is replaced by a " $ " sign and whenever we need the datetime of the image, we replace " $ " by " : "  sing. there are other ways to do this but i went with this one